remove wallet unlocked and wallet locked
