+++ e2e/tests/aws.rs
@@ -2,7 +2,8 @@
 mod tests {
     use anyhow::Result;
     use e2e::e2e_helpers::start_aws_kms;
-    use fuels::accounts::kms::AwsWallet;
+    use fuels::accounts::kms::KmsKey;
+    use fuels::accounts::wallet::NewWallet;
     use fuels::accounts::{Account, ViewOnlyAccount};
     use fuels::prelude::{
         launch_provider_and_get_wallet, AssetId, Contract, LoadConfiguration, TxPolicies,
@@ -27,7 +28,8 @@ mod tests {
         let provider = wallet.provider().clone();
 
         // ANCHOR: use_kms_wallet
-        let wallet = AwsWallet::with_kms_key(your_kms_key_id, kms.client(), Some(provider)).await?;
+        let kms_key = KmsKey::new(your_kms_key_id, kms.client()).await?;
+        let wallet = NewWallet::new(kms_key, provider);
         // ANCHOR_END: use_kms_wallet
 
         let total_base_balance = wallet.get_asset_balance(&AssetId::zeroed()).await?;
@@ -53,14 +55,14 @@ mod tests {
         let your_kms_key_id = key.id;
         let provider = wallet.provider().clone();
 
-        let aws_wallet =
-            &AwsWallet::with_kms_key(your_kms_key_id, kms.client(), Some(provider)).await?;
+        let kms_key = KmsKey::new(your_kms_key_id, kms.client()).await?;
+        let aws_wallet = NewWallet::new(kms_key, provider);
 
         Contract::load_from(
             "../e2e/sway/contracts/contract_test/out/release/contract_test.bin",
             LoadConfiguration::default(),
         )?
-        .deploy(aws_wallet, TxPolicies::default())
+        .deploy(&aws_wallet, TxPolicies::default())
         .await?;
 
         Ok(())
